[db] Add an "address" column to the builders table
This column is meant to store the address of the buildd software in the
remote machine, to which eg. ACCEPTED mails from dak should be sent to.

It should be stored in the database so that there is no duplication of
information between the DB (list of builders) and our aliases file (list
of builders and their addresses).

Signed-off-by: Roger Leigh <rleigh@debian.org>