[db] Distinguish between arches with buildds, and the Architecture field
Previously, the architectures table was holding real architectures like
"alpha" and "mipsel", and other values like "any" and "all". However, the
architectures table is referenced from places like build_jobs and builder,
where only the former group makes sense.

This commit splits the two groups, keeping the architectures table for the
real architectures (or, rather, architectures supported by each wanna-build
instance), and creates a new table package_architectures for values found in
the Architecture field of Sources/Packages files.

Additionally, a trigger is created to populate package_architectures upon
insertion on source_architectures and binaries. It is not feasible to
populate this table in advance: at the time of this writing, there are 38
different names in the Architecture field in Sources.gz, with more appearing
in an unforeseeable manner.

Signed-off-by: Roger Leigh <rleigh@debian.org>