[db] Prefer column names derived from table name, and drop _name suffix
Instead of using a column "name" in every table, use a name derived from
the table name, eg. "source" in sources and "arch" in architectures. Also,
when creating references, don't name such columns with a "_name" suffix;
instead, just use the name of the referenced column if possible. (It is not
possible, for example, in binaries, where there are two version numbers.)

With this, it is possible to create queries more easily, and natural joins
are possible.

Incidentally, this commits removes the long listing of package section,
which is unnecessary given that they will be auto-inserted via a trigger
when populating the sources and binaries table.

Signed-off-by: Roger Leigh <rleigh@debian.org>