[db] Fix the primary key in the suite_binaries table
The primary key in "suite_binaries" was a composite key of (pacakge,
suite), but it is obvious that (package, suite) has to be duplicated
once per arch: add "arch" to the primary key.

Additionally, the UNIQUE constraint is unnecessary: with (package, arch,
suite) being unique already due to the primary key, there'll be only one
version for each combination.

Signed-off-by: Roger Leigh <rleigh@debian.org>