[db] Create a "suite_arches" table to list the architectures in a suite
The database schema was missing a list of architectures in each suite.
This is needed, because one can't simply use the "architectures" table
to decide what arches to act on, eg. when importing packages
information, or when creating a web interface: that table may contain no
longer active architectures (which are kept around for build_logs), and
many actions really need the list of architectures specific to a suite.

This commit introduces a "suite_arches" table, with a relation of what
architectures are available in each suite. It also adds foreign keys
towards this table from suite_binaries and build_status; this way, it is
ensured that dropping an architecture from a suite will clean up those
tables from no longer relevant entries.

Signed-off-by: Roger Leigh <rleigh@debian.org>