[buildd] Remove apt cache updating through host apt-get
Using the host apt-get only works if the binary format of the cache has
not changed and especially not if the host architecture differs from
the one of the chroot.  sbuild will run apt-get update if told to do so
before building anyway so we could just delete the lines responsible
for this.

Signed-off-by: Philipp Kern <pkern@debian.org>