Temporarily install a nssdatabases copy script for schroot.
schroot will already provide a means to copy dumps of the NSS databases
into the chroot in its next revision.  Sadly this is not available yet
for use on the d.o boxes and so we resort on a temporary copy of the
script with configuration in `/etc/sbuild'.

We should take care of applying http://www.dpkg.org/dpkg/ConffileHandling
properly to get rid of this as soon as we can depend on a backported
newer version of sbuild.

Signed-off-by: Philipp Kern <pkern@debian.org>