[Sbuild::LogBase] Don't close STDOUT or STDERR
When reopening, we can't guarantee getting the same fd (1 or 2).
Therefore just reopen in place.

Redirect whether or not in verbose mode (since we might want to log
what's printed to STDOUT or STDERR, and the real original fds are
still open in the logger coprocess; output to STDOUT or STDERR in the
parent will therefore be redirected to the logger and logged both to
file and/or standard output depending upon the configuration.