[*::Conf] Use eval in place of require
Using require means that %INC has to be messed around with in order to
read the same configuration file more than once.  It also means that by
default any invalid variables in the configuration file are not
reported.

Also move wanna-build configuration back to wanna-build.conf, but using
the new syntax.