[build] Package versioning uses git release tags
Rather than hard-coding the release version in several places,
including configure.ac and NEWS as well as the git tags, make the
setup simpler and easier to manage for releases since forgetting
to update one of the versions or tag is not possible.

The version is stored in the NEWS file.  This is used to tag the
release with scripts/git-tag-release which creates a tag
release/sbuild-<version>.

The script ./bootstrap calls the script scripts/git-version which
gets the release tag using git and uses this to create a VERSION
file containing the release details.  The autotools read this
file to get the version to set in the configure script.  The
reason for this is to allow rebootstrapping from release tarballs
without needing git.

Debian releases are tagged with debian/git-tag-debian.

The release procedures are documented in HACKING.