buildd: Make multi-archive-capable
Allow buildds to use a new config format allowing for building
packages from several archives. For this, we keep a distribution
configuration around and pass this to all methods that are doing
dist-specific things.

An important change is that there is no central 'DB' handle for
contacting wanna-build anymore, but that it is created on request,
depending on the distribution configuration.

Old configurations are transparently converted to the new format.