Sbuild::Chroot: Simplify logic for executing on host
Previously, Chroot classes had a CHROOT=>0|1 parameter
to control whether the command was executed inside the
chroot or on the host system.  This has been removed
in favour of using Sbuild::ChrootRoot for executing on
the host.  Note that user switching on the host is not
permitted.  This means that root access is only allowed
in the build chroot.