sbuild: remove EXIT-DAEMON-PLEASE handling
EXIT-DAEMON-PLEASE is an implementation detail of buildd-watcher.
buildd does not call sbuild with multiple packages anymore.  It might
be sane for batch mode to get a similar flag file back eventually,
but currently it even breaks wanna-build signaling because sbuild does
not deliver the proper exit code back to the caller.