Sbuild::InternalBuildDepSatisfier: Do not try other alternatives if installing default version is enough
When we encounter

  Build-Depends: pkg-a (>= 2) | pkg-b (>= 3)

and a lower version of pkg-a is already installed (or provided by
another package), sbuild would try to upgrade pkg-b as well.

This patch changes the behaviour of sbuild to not try satisfying
alternative dependencies if the dependency can be satisfied by
installing the default version.

Bug-Debian: http://bugs.debian.org/586275
Signed-off-by: Ansgar Burchardt <ansgar@43-1.org>
Signed-off-by: Roger Leigh <rleigh@debian.org>