Sbuild::Build: Remove support for stderr parsing in run_apt
STDERR wasn't being piped back by pipe_apt_command for some time,
so this code was not actually being used.
Diagnosing causes of failure in the build environment was pointless,
since it was logged anyway, and if the messages changed would be
missed.
Virtual dependency resolving using the static ALTERNATIVES list has
been removed.  This was also broken, and the aptitude resolver can
do a proper job if this is required.