Sbuild::InternalBuildDepSatisfier: Resolve simple virtual dependencies
get_virtual gets a list of packages providing a given package
(or a single concrete package if it's not virtual), using
"apt-cache --names-only search '^$pkg$'".  This appears, at
present, to be the most efficient way of finding virtual
package providers.  Packages are listed in alphabetical order.

virtual_dependencies uses get_virtual to expand all virtual
dependencies.  It currently picks the first returned package,
which is the first sorted alphabetically.