etc/99builddsourceslist: rip out support for key packages
It doesn't seem to be possible to use apt-get via simple chroot
statements in a sane way from this script.  You'd need to update
its cache first because the new sources.list might not refer to
any mirrors whose lists are already cached and then dpkg complains
about its environment not being set up properly.

So let's add the debian-edu key to the package and handle it like
the backports.org key.