Sbuild::Chroot: Filter environment using ENVIRONMENT_FILTER
Sbuild::Conf: Add ENVIRONMENT_FILTER/$environment_filter option
The environment filter is a list of regexes.  If an environment
variable matches any of the regexes, it is kept, or else if none
match it it dropped.  Currently various DEB* and *FLAGS* are
allowed.

The allowed regexes are taken from debuild.  debuild also allows
LC_* variables, but sbuild does not (it sets the locale to POSIX,
making this pointless to allow).