sbuild: Simplify shutdown logic
Remove unused and redundant REDO/SBUILD-REDO-DUMPED code.  This hasn't
been used for 11 months (explicitly disabled for buildds).  buildd and
buildd-mail continue to read and write  REDO for their own internal
uses.

Shutdown sequence is aligned with the build sequence in Sbuild::Build.

Temporary variables for job, session and resolver are used to simplify
code.

The same error return is issued when in buildd or user mode and killed
by a signal.