Sbuild::Conf: Update and upgrade the build chroot by default
Having an up-to-date chroot prior to installing any build dependencies
means it's less likely to break the chroot when removing the packages
post-build.  This is because we're starting and ending at the same
baseline state.  If the chroot is out of date, then it becomes possible
to break the package manager because packages to reinstall are no
longer available, or there are version mismatches etc., which require
manual intervention to get back to a clean state.