Sbuild::ResolverBase: Implement alternative stripping
The apt and aptitude resolvers now only use the first alternative
in alternative dependencies.  This does not include architecture-
specific alternatives, which are reduced prior to stripping.

This feature is configurable ($resolve_alternatives), defaulting
to being enabled.  For suites which can take advantage of
alterntatives (e.g. backports), this may be disabled.  For unstable,
where we want the strict consistency and reproducibility guaranteed
by stripping, this makes the apt and aptitude resolvers behave
exactly like the internal resolver.