Allow MIME encoded build logs.
If $mime_build_log_mails is set to 1, then you will get a multipart/mixed mail
comprising either one or two parts: the build log is always included as an
attached file, the .changes file is attached only if the build completed
successfully.  If $compress_build_log_mails is set to 1, then the build log
will be sent gzip-encoded instead of plain text.

Signed-off-by: Philipp Kern <pkern@debian.org>