Sbuild::Build: Add exceptions to run()
run() currently uses a number of gotos and labels to clean up on
error.  Unfortunately, this logic is both complex and fragile;
cleanup tasks need doing in the correct order, or not at all.
run() has been split up into separate functions to make the nested
logic explicit, and now throws exceptions on failure; cleanup can
be done before or inside the exception handlers, and the exceptions
may be rethrown to allow cleanup back up the call stack.