sbuild: Use sbuild user for building
Rather than using the user invoking sbuild, everything inside the
chroot is done either as root or sbuild.  The intention behind this
is to increase the privilege separation between the user and sbuild,
with the ultimate aim being the user having no access to the build
chroot; all access will be mediated via sbuild.  The current change
has several implications:

- a wrapper program, sbuild-schroot is used to run schroot.  This
  handles switching to the sbuild user inside the chroot.
- all files created inside the chroot must be owned and writable
  by user and group sbuild.  The user ownership is required so that
  the sbuild user can create and delete files.  The group ownership
  is so that the user can create and delete files.
- group write access is removed during the build stage, to prevent
  external tampering.

Note that due to the architecture of sbuild, the user may still
interfere with the process themselves, and gain root inside the
chroot either via schroot or sbuild-schroot.  These loopholes will
be closed by future changes.  The current changes do prevent
accidental tampering, but not deliberate interference.