sbuild: Add safe signal handling
Remove all cleanup logic from the shutdown signal handler.  The
handler now just calls the abort method on the Build object, which
sets an ABORT variable containing the reason for aborting.  The build
sequence calls the check_abort method to poll for an ABORT request at
frequent intervals (before every long running job, and for each log
line printed when dpkg-buildpackage is being run.

The normal cleanup and shutdown logic is used on trapping a signal,
so the chroot should always be left clean.  Once exception is if hitting
C-c from a shell, which may cause SIGINT to be sent to all processes in
the process group when apt-get or aptitude are part way through
installing packages.  This may or may not result in a bad package state
inside the chroot (but is more reliable than the existing method).
Signals may now also be received multiple times, leading to multiple
ABORT requests, but not premature program termination.