Sbuild::Build: Improve binNMU changelog handling
According to https://wiki.ubuntu.com/MultiarchSpec#Binary%20NMUs ,
although a Multi-Arch: same package may not be binNMU’d on individual
architectures, it should be possible to schedule a binNMU on all
architectures at once, and have the resulting packages remain
co-installable.  However, this does not quite work right now for a silly
reason: sbuild generates a different timestamp for the binNMU changelog
entry in each build, and the differing changelog.gz files break
co-installability.

This fixes the problem by generating the timestamp in a deterministic way:
we just copy the timestamp from the previous changelog entry.  (Since the
source has not changed since that time, it seems like as good a time as
any to write into the changelog.)

Certain automated backport setups that use --append-to-version are also
affected by this issue and require this fix as well.

Use dpkg-parsechangelog to parse previous changelog entry to acquire
all needed information, including the correct date.

Signed-off-by: Roger Leigh <rleigh@debian.org>