Sbuild::ResolverBase: Allow alternative dependencies for same package
Currently, when alternative dependencies are not allowed, only the
first alternative is used.  Here, this rule is relaxed when the
alternative dependency or dependencies are for the same package name,
i.e. foo (rel x) | foo (rel y) [ ... foo (rel z) ]
since while these are separate dependencies, they are in reality a
single logical dependency.