Sbuild::Build: Only log to one logfile, dropping build log
Only create the package build log, dropping the main build log
(which would only ever be used to log errors prior to a build
commencing, following which all output would be redirected to the
package build log).

This removes a layer of I/O redirection, which complicated the
logging code, and added an extra layer of hairy complexity.