build: Install /var/lib/sbuild/build
Add directory for holding build trees; this can be bind
mounted by schroot when starting a new session.  This
makes it easier to get at builds when using snapshot
chroots, and also means the space for the build doesn't
need to be in the chroot when using LVM snapshots, which
would otherwise waste disc space.