sbuild: Move http downloads to executable wrapper
This permits the .dsc to not match the standard naming
conventions providing that the contents are correct.  The
main goal is to only have the core code accept .dsc or
package_version for download using apt-get.  This is a
prerequisite for proper privilege separation and
simplification of the core code.